# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Stellar wallet backend — auth, encrypted credential backup, and account recovery.",
        "title": "Latch API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/smart-account/balances": {
            "get": {
                "description": "Fetches balances for every asset in the configured catalog. Zero balances are omitted unless all=1.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "smart-account"
                ],
                "summary": "Get a smart account's asset balances",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Smart account address",
                        "name": "smartAccountAddress",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Include zero balances when set to 1",
                        "name": "all",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler_webapp.webappErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler_webapp.webappErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/smart-account/context-rules": {
            "get": {
                "description": "Fetches the on-chain context rules (spending policies and their signers) configured for a smart account.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "smart-account"
                ],
                "summary": "List a smart account's context rules",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Smart account address",
                        "name": "address",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "testnet",
                        "description": "Network",
                        "name": "network",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler_webapp.webappErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler_webapp.webappErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/smart-account/webauthn": {
            "get": {
                "description": "Pure computation over client-supplied key material (no session, no persistence). Returns the deterministic smart account address for the given credential/key data and whether it's already deployed.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "smart-account"
                ],
                "summary": "Derive a smart account address from WebAuthn key material",
                "parameters": [
                    {
                        "type": "string",
                        "description": "WebAuthn credential ID",
                        "name": "credentialId",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Hex-encoded public key + credential ID",
                        "name": "keyDataHex",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler_webapp.webappErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler_webapp.webappErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Standalone deploy endpoint over client-supplied key material — does not persist a webapp.smart_accounts row (that only happens via the registration-finish flow, which is tied to a session user).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "smart-account"
                ],
                "summary": "Deploy a smart account from WebAuthn key material",
                "parameters": [
                    {
                        "description": "WebAuthn key data and credential ID",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler_webapp.deploySmartAccountRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler_webapp.webappErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler_webapp.webappErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_handler_webapp.deploySmartAccountRequest": {
            "type": "object",
            "required": [
                "credentialId",
                "keyDataHex"
            ],
            "properties": {
                "credentialId": {
                    "type": "string"
                },
                "keyDataHex": {
                    "type": "string"
                }
            }
        },
...
        "internal_handler_webapp.webappErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "internal_error"
                },
                "error": {
                    "type": "string",
                    "example": "internal error"
                },
                "message": {
                    "type": "string",
                    "example": "internal error"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter: Bearer \u003caccess_token\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "RecoveryAuth": {
            "description": "Enter: Bearer \u003crecovery_token\u003e  (obtained from POST /v1/recovery/verify)",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}